requirements.txt
pandas
numpy
scikit-learn
plotly
shinywidgets
scipy
